Inline resolve_signature helper, merge tests
Move resolve_signature into jit_argument.py (both callsites already
import from there) instead of a standalone _annotations.py file, and
fold the future-annotations test into the existing test_jit_cache_key.py
which it naturally extends.
