+++ tests/unit/test_jit_cache_key.py
@@ -1,6 +1,7 @@
+from __future__ import annotations
+
 # SPDX-License-Identifier: Apache-2.0
 # Copyright (c) 2025 FlyDSL Project Contributors
-
 import flydsl.compiler as flyc
 import flydsl.expr as fx
 from flydsl.compiler.jit_argument import JitArgumentRegistry
@@ -20,6 +21,11 @@ def _stream_launch(stream: fx.Stream = fx.Stream(None)):
 
 @flyc.jit
 def _constexpr_launch(value: fx.Constexpr[int]):
+    pass
+
+
+@flyc.jit
+def _runtime_int32_launch(n: fx.Int32, stream: fx.Stream = fx.Stream(None)):
     pass
 
 
@@ -45,3 +51,13 @@ def test_stream_cache_key_ignores_runtime_representation():
 
 def test_constexpr_values_still_participate_in_cache_key():
     assert _cache_key(_constexpr_launch, 1) != _cache_key(_constexpr_launch, 2)
+
+
+def test_future_annotations_runtime_int32_ignores_value_in_cache_key():
+    """`from __future__ import annotations` stringifies fx.Int32; resolve_signature must eval it back so the value stays out of the cache key."""
+    key1 = _cache_key(_runtime_int32_launch, 1)
+    key2 = _cache_key(_runtime_int32_launch, 2)
+
+    assert key1 == key2
+    assert ("n", int) in key1
+    assert ("n", (int, 1)) not in key1
